+++ fnos/__init__.py
@@ -31,6 +31,6 @@ from .share import Share
 from .notify import Notify
 from .iscsi_manager import IscsiManager
 
-__version__ = "0.11.0"
+__version__ = "0.12.0"
 
 __all__ = ["FnosClient", "Store", "ResourceMonitor", "SAC", "SystemInfo", "User", "Network", "File", "DockerManager", "EventLogger", "Share", "Notify", "IscsiManager"]